Moving the tests around and updating the setup.py
