--- test/test_env.py
@@ -1,35 +1,0 @@
-#!/usr/bin/env python
-
-import os
-import unittest
-from yarn.api import env
-
-
-class TestEnv(unittest.TestCase):
-
-    def test_env_port(self):
-        env.host_port = 66
-        assert env._port == 66 and env.host_port == 66
-
-    def test_env_port_non_int(self):
-        def set_port_non_int():
-            env.host_port = "test"
-        self.assertRaises(AttributeError, set_port_non_int)
-
-    def test_env_port_bad_int(self):
-        def set_port_bad_int():
-            env.host_port = 111111
-        self.assertRaises(AttributeError, set_port_bad_int)
-
-    def test_env_user_autoset(self):
-        assert env.user == os.environ["USERNAME"]
-
-    def test_env_username_bad_input(self):
-        def set_bad_username():
-            env.user = 12
-        self.assertRaises(AttributeError, set_bad_username)
-
-    def test_env_password_bad_input(self):
-        def set_bad_password():
-            env.password = 12
-        self.assertRaises(AttributeError, set_bad_password)
